+++ doc/changelog.md
@@ -3,6 +3,8 @@
 * **0.27-SNAPSHOT**
   - Update to JMockit 1.43
   - Compiles with Java 11
+  - Update to jnr-unixsocket version to 0.21 (#1089)
+  - Add 'readOnly' option for docker:run cto mount container's root fs read-only (#1125)
 
 * **0.27.2** (2018-10-05)